feat : lecture에서 course로 변경
