+++ src/main/java/sopio/acha/domain/activity/infrastructure/ActivityRepository.java
@@ -12,6 +12,6 @@ import sopio.acha.domain.activity.domain.Activity;
 public interface ActivityRepository extends JpaRepository<Activity, Long> {
 	boolean existsActivityByTitleAndMemberId(String title, String memberId);
 	List<Activity> findTop10ByMemberIdAndDeadlineAfterOrderByDeadlineAsc(String memberId, LocalDateTime now);
-	List<Activity> findAllByMemberIdAndLectureIdOrderByWeekAsc(String memberId, Long lectureId);
+	List<Activity> findAllByMemberIdAndCourseIdOrderByWeekAsc(String memberId, Long CourseId);
 	List<Activity> findAllByDeadlineAfterAndNotifyScheduledIsFalse(LocalDateTime now);
 }
